+++ docs/demo/demo-editable.js
@@ -16,7 +16,7 @@ new mar10.Wunderbaum({
   id: "demo",
   element: document.getElementById("demo-tree"),
   debugLevel: 5,
-  connectTopBreadcrumb: document.getElementById("parentPath"),
+  connectTopBreadcrumb: "output#parentPath",
   checkbox: true,
   // fixedCol: true,
   // navigationModeOption: "row",