+++ CHANGELOG.md
@@ -7,6 +7,8 @@ First release.
 > This section will be removed after the beta phase. <br>
 > Note that semantic versioning rules are not strictly followed during this phase.
 
+- feature/breadcrumb-improvements: connectTopBreadcrumb can be a selector
+
 - v0.13.0: Add `tree.countUnique()` method.
 
 - v0.12.1: Fix flat source format for positional args.